# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  helpers.Response:
    properties:
      data: {}
      error: {}
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  permissions.CheckPermissions:
    properties:
      method:
        type: string
      path:
        type: string
    type: object
  permissions.CreatePermissions:
    properties:
      method:
        type: string
      p_type:
        type: string
      route:
        type: string
      subject:
        type: string
    type: object
  permissions.DeletePermissions:
    properties:
      method:
        type: string
      p_type:
        type: string
      route:
        type: string
      subject:
        type: string
    type: object
  permissions.EditPermissions:
    properties:
      new_policy:
        $ref: '#/definitions/permissions.Policy'
      old_policy:
        $ref: '#/definitions/permissions.Policy'
    type: object
  permissions.Policy:
    properties:
      method:
        type: string
      p_type:
        type: string
      route:
        type: string
      subject:
        type: string
    type: object
...
info:
  contact: {}
  description: Go API service
  title: Go API
  version: "1.0"
paths:
...
  /permissions/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/permissions.CheckPermissions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Check permission
      tags:
      - permissions
  /permissions/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/permissions.CreatePermissions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Create RBAC policy
      tags:
      - permissions
  /permissions/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/permissions.DeletePermissions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Delete RBAC policy
      tags:
      - permissions
  /permissions/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/permissions.EditPermissions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Update RBAC policy
      tags:
      - permissions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by the access token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"